# swagger/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "LazerVault API",
    "description": "LazerVault Account Card API",
    "version": "1.0"
  },
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/face/register": {
      "post": {
        "summary": "Register Face",
        "description": "Register a new face for a user with optional duplicate detection. Supports JPG/PNG images up to 32MB.",
        "operationId": "FacialRecognitionService_RegisterFace",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbRegisterFaceResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": "Request to register a new face for facial recognition",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbRegisterFaceRequest"
            }
          }
        ],
        "tags": [
          "Facial Recognition"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/v1/face/verify": {
      "post": {
        "summary": "Verify Face",
        "description": "Verify a user's identity by comparing their face against registered faces. Returns confidence score and verification result.",
        "operationId": "FacialRecognitionService_VerifyFace",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbVerifyFaceResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": "Request to verify a user's face against registered faces",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbVerifyFaceRequest"
            }
          }
        ],
        "tags": [
          "Facial Recognition"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "pbDuplicateDetails": {
      "type": "object",
      "properties": {
        "isDuplicate": {
          "type": "boolean"
        },
        "threshold": {
          "type": "number",
          "format": "double"
        },
        "totalMatches": {
          "type": "integer",
          "format": "int32"
        },
        "primaryMatch": {
          "$ref": "#/definitions/pbPrimaryMatch"
        },
        "allMatches": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/pbMatch"
          }
        },
        "message": {
          "type": "string"
        },
        "securityNote": {
          "type": "string"
        }
      },
      "title": "Duplicate Details for Registration"
    },
...
    "pbMatch": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string"
        },
        "faceId": {
          "type": "string"
        },
        "confidence": {
          "type": "number",
          "format": "double"
        },
        "registeredAt": {
          "type": "string"
        }
      },
      "title": "Match Details"
    },
...
    "pbPrimaryMatch": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string"
        },
        "faceId": {
          "type": "string"
        },
        "confidence": {
          "type": "number",
          "format": "double"
        },
        "registeredAt": {
          "type": "string"
        }
      },
      "title": "Primary Match Details"
    },
...
    "pbRegisterFaceRequest": {
      "type": "object",
      "properties": {
        "user_id": {
          "type": "string"
        },
        "face_id": {
          "type": "string"
        },
        "allow_duplicates": {
          "type": "boolean"
        },
        "duplicate_threshold": {
          "type": "number",
          "format": "double"
        },
        "image_data": {
          "type": "string",
          "format": "byte"
        },
        "image_filename": {
          "type": "string"
        },
        "image_content_type": {
          "type": "string"
        }
      },
      "description": "Request to register a new face for facial recognition",
      "title": "Register Face Request",
      "required": [
        "user_id",
        "image_data"
      ]
    },
    "pbRegisterFaceResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "faceId": {
          "type": "string"
        },
        "message": {
          "type": "string"
        },
        "error": {
          "type": "string"
        },
        "numFacesDetected": {
          "type": "integer",
          "format": "int32"
        },
        "duplicateDetails": {
          "$ref": "#/definitions/pbDuplicateDetails"
        }
      },
      "title": "Register Face Response"
    },
...
    "pbVerifyFaceRequest": {
      "type": "object",
      "properties": {
        "user_id": {
          "type": "string"
        },
        "threshold": {
          "type": "number",
          "format": "double"
        },
        "image_data": {
          "type": "string",
          "format": "byte"
        },
        "image_filename": {
          "type": "string"
        },
        "image_content_type": {
          "type": "string"
        }
      },
      "description": "Request to verify a user's face against registered faces",
      "title": "Verify Face Request",
      "required": [
        "user_id",
        "image_data"
      ]
    },
    "pbVerifyFaceResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "verified": {
          "type": "boolean"
        },
        "confidence": {
          "type": "number",
          "format": "double"
        },
        "matchedFaceId": {
          "type": "string"
        },
        "threshold": {
          "type": "number",
          "format": "double"
        },
        "distance": {
          "type": "number",
          "format": "double"
        },
        "message": {
          "type": "string"
        },
        "error": {
          "type": "string"
        }
      },
      "title": "Verify Face Response"
    },
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}